Refactor Image Options and fix thresholdCache
