+++ packages/core/src/web/app/components/beambox/RightPanel/OptionsBlocks/ImageOptions/ImageOptions.module.scss
@@ -1,4 +1,4 @@
-@use '../_mixins.scss' as mixins;
+@use '../../_mixins.scss' as mixins;
 
 .field {
   display: flex;
